Changes to be committed: 	new file:   odometry/CMakeLists.txt new file:   odometry/package.xml 	new file:   odometry/src/ticks_odom.py deleted:    odometry/ticks_odom.py
 Changes not staged for commit:
	modified:   kinematics_pkg/CMakeLists.txt
	modified:   kinematics_pkg/package.xml
